Bug affichage enfin fixed

+++ src/render/raycasting.c
@@ -19,8 +19,7 @@ void	cast_rays(t_game *game)
 
 	ray = game->ray;
 	i = 0;
-	ray->angle = game->player->angle - (game->player->fov / 2.0);
-	// printf("ANCIEN : %f, NOUVEAU : %f\n", M_PI / 6, game->player->fov / 2.0);
+	ray->angle = game->player->angle - (game->player->fov / 2);
 	while (i < WIN_WIDTH)
 	{
 		init_ray(ray);
@@ -28,7 +27,7 @@ void	cast_rays(t_game *game)
 		ray->wall_dist = ray->wall_dist * cos(game->player->angle - ray->angle);
 		render_wall(game, ray, i);
 		render_floor_ceiling(game, ray, i);
-		ray->angle += game->player->fov / (double)WIN_WIDTH;
+		ray->angle += game->player->fov / WIN_WIDTH;
 		if (ray->angle > 2 * M_PI)
 			ray->angle -= 2 * M_PI;
 		i++;
